chore: move the 'more_like_this' functions to the pdb schema
