+++ pg_search/src/api/builder_fns/mod.rs
@@ -15,6 +15,7 @@
 // You should have received a copy of the GNU Affero General Public License
 // along with this program. If not, see <http://www.gnu.org/licenses/>.
 
+mod mlt;
 mod paradedb;
 mod pdb;
 mod proximity;
