+++ tests/tests/mlt.rs
@@ -27,7 +27,7 @@ fn mlt_enables_scoring_issue1747(mut conn: PgConnection) {
     SimpleProductsTable::setup().execute(&mut conn);
 
     let (id,) = "
-    SELECT id FROM paradedb.bm25_search WHERE id @@@ paradedb.more_like_this(
+    SELECT id FROM paradedb.bm25_search WHERE id @@@ pdb.more_like_this(
         document_id => 3,
         min_term_frequency => 1
     ) ORDER BY id LIMIT 1"
@@ -44,7 +44,7 @@ fn mlt_scoring_nested(mut conn: PgConnection) {
     SELECT * FROM paradedb.bm25_search
     WHERE id @@@ 
     paradedb.boolean(
-        must => paradedb.more_like_this(
+        must => pdb.more_like_this(
             min_doc_frequency => 2,
             min_term_frequency => 1,
             document_fields => '{"description": "keyboard"}'
@@ -60,7 +60,7 @@ fn mlt_scoring_nested(mut conn: PgConnection) {
     SELECT * FROM paradedb.bm25_search
     WHERE id @@@ 
     paradedb.boolean(
-        must_not => paradedb.more_like_this(
+        must_not => pdb.more_like_this(
             min_doc_frequency => 2,
             min_term_frequency => 1,
             document_fields => '{"description": "keyboard"}'
@@ -76,7 +76,7 @@ fn mlt_scoring_nested(mut conn: PgConnection) {
     SELECT * FROM paradedb.bm25_search
     WHERE id @@@ 
     paradedb.boolean(
-        should => paradedb.more_like_this(
+        should => pdb.more_like_this(
             min_doc_frequency => 2,
             min_term_frequency => 1,
             document_fields => '{"description": "keyboard"}'
@@ -93,7 +93,7 @@ fn mlt_scoring_nested(mut conn: PgConnection) {
     WHERE id @@@ 
     paradedb.boost(
         factor => 1.5,
-        query => paradedb.more_like_this(
+        query => pdb.more_like_this(
             min_doc_frequency => 2,
             min_term_frequency => 1,
             document_fields => '{"description": "keyboard"}'
@@ -110,7 +110,7 @@ fn mlt_scoring_nested(mut conn: PgConnection) {
     WHERE id @@@ 
     paradedb.const_score(
         score => 5,
-        query => paradedb.more_like_this(
+        query => pdb.more_like_this(
             min_doc_frequency => 2,
             min_term_frequency => 1,
             document_fields => '{"description": "keyboard"}'
@@ -127,12 +127,12 @@ fn mlt_scoring_nested(mut conn: PgConnection) {
     WHERE id @@@ 
     paradedb.disjunction_max(
         disjuncts => ARRAY[
-            paradedb.more_like_this(
+            pdb.more_like_this(
                 min_doc_frequency => 2,
                 min_term_frequency => 1,
                 document_fields => '{"description": "keyboard"}'
             ), 
-            paradedb.more_like_this(
+            pdb.more_like_this(
                 min_doc_frequency => 2,
                 min_term_frequency => 1,
                 document_fields => '{"description": "shoes"}'
@@ -154,13 +154,13 @@ fn mlt_scoring_nested(mut conn: PgConnection) {
             disjuncts => ARRAY[
                 paradedb.boost(
                     factor => 3,
-                    query => paradedb.more_like_this(
+                    query => pdb.more_like_this(
                         min_doc_frequency => 2,
                         min_term_frequency => 1,
                         document_fields => '{"description": "keyboard"}'
                     ) 
                 ),
-                paradedb.more_like_this(
+                pdb.more_like_this(
                     min_doc_frequency => 2,
                     min_term_frequency => 1,
                     document_fields => '{"description": "shoes"}'
