+++ tests/tests/query.rs
@@ -385,7 +385,7 @@ fn more_like_this_raw(mut conn: PgConnection) {
 
     match r#"
     SELECT id, flavour FROM test_more_like_this_table WHERE test_more_like_this_table @@@
-        paradedb.more_like_this();
+        pdb.more_like_this();
     "#
     .fetch_result::<()>(&mut conn)
     {
@@ -399,7 +399,7 @@ fn more_like_this_raw(mut conn: PgConnection) {
 
     let rows: Vec<(i32, String)> = r#"
     SELECT id, flavour FROM test_more_like_this_table
-    WHERE test_more_like_this_table @@@ paradedb.more_like_this(
+    WHERE test_more_like_this_table @@@ pdb.more_like_this(
         min_doc_frequency => 0,
         min_term_frequency => 0,
         document_fields => '{"flavour": "banana"}'
@@ -410,7 +410,7 @@ fn more_like_this_raw(mut conn: PgConnection) {
 
     let rows: Vec<(i32, String)> = r#"
     SELECT id, flavour FROM test_more_like_this_table
-    WHERE test_more_like_this_table @@@ paradedb.more_like_this(
+    WHERE test_more_like_this_table @@@ pdb.more_like_this(
         min_doc_frequency => 0,
         min_term_frequency => 0,
         document_id => 2
@@ -444,7 +444,7 @@ fn more_like_this_empty(mut conn: PgConnection) {
 
     match r#"
     SELECT id, flavour FROM test_more_like_this_table
-    WHERE test_more_like_this_table @@@ paradedb.more_like_this()
+    WHERE test_more_like_this_table @@@ pdb.more_like_this()
     ORDER BY id;
     "#
     .fetch_result::<()>(&mut conn)
@@ -482,7 +482,7 @@ fn more_like_this_text(mut conn: PgConnection) {
 
     let rows: Vec<(i32, String)> = r#"
     SELECT id, flavour FROM test_more_like_this_table
-    WHERE test_more_like_this_table @@@ paradedb.more_like_this(
+    WHERE test_more_like_this_table @@@ pdb.more_like_this(
         min_doc_frequency => 0,
         min_term_frequency => 0,
         document_fields => '{"flavour": "banana"}'
@@ -515,7 +515,7 @@ fn more_like_this_boolean_key(mut conn: PgConnection) {
     let rows: Vec<(bool, String)> = r#"
     SELECT id, flavour FROM test_more_like_this_table
     WHERE test_more_like_this_table @@@
-    paradedb.more_like_this(
+    pdb.more_like_this(
        min_doc_frequency => 0,
        min_term_frequency => 0,
        document_fields => '{"flavour": "banana"}'
@@ -549,7 +549,7 @@ fn more_like_this_uuid_key(mut conn: PgConnection) {
 
     let rows: Vec<(uuid::Uuid, String)> = r#"
     SELECT id, flavour FROM test_more_like_this_table
-    WHERE test_more_like_this_table @@@ paradedb.more_like_this(
+    WHERE test_more_like_this_table @@@ pdb.more_like_this(
         min_doc_frequency => 0,
         min_term_frequency => 0,
         document_fields => '{"flavour": "banana"}'
@@ -584,7 +584,7 @@ fn more_like_this_i64_key(mut conn: PgConnection) {
     let rows: Vec<(i64, String)> = r#"
     SELECT id, flavour FROM test_more_like_this_table
     WHERE test_more_like_this_table @@@
-    paradedb.more_like_this(
+    pdb.more_like_this(
         min_doc_frequency => 0,
         min_term_frequency => 0,
         document_fields => '{"flavour": "banana"}'
@@ -618,7 +618,7 @@ fn more_like_this_i32_key(mut conn: PgConnection) {
 
     let rows: Vec<(i32, String)> = r#"
     SELECT id, flavour FROM test_more_like_this_table
-    WHERE test_more_like_this_table @@@ paradedb.more_like_this(
+    WHERE test_more_like_this_table @@@ pdb.more_like_this(
         min_doc_frequency => 0,
         min_term_frequency => 0,
         document_fields => '{"flavour": "banana"}'
@@ -652,7 +652,7 @@ fn more_like_this_literal_cast(mut conn: PgConnection) {
 
     let rows: Vec<(i32, i32)> = r#"
     SELECT id, year FROM test_more_like_this_table
-    WHERE test_more_like_this_table @@@ paradedb.more_like_this(
+    WHERE test_more_like_this_table @@@ pdb.more_like_this(
         min_doc_frequency => 0,
         min_term_frequency => 0,
         document_fields => '{"year": 2012}'
@@ -686,7 +686,7 @@ fn more_like_this_i16_key(mut conn: PgConnection) {
 
     let rows: Vec<(i16, String)> = r#"
     SELECT id, flavour FROM test_more_like_this_table
-    WHERE test_more_like_this_table @@@ paradedb.more_like_this(
+    WHERE test_more_like_this_table @@@ pdb.more_like_this(
         min_doc_frequency => 0,
         min_term_frequency => 0,
         document_fields => '{"flavour": "banana"}'
@@ -720,7 +720,7 @@ fn more_like_this_f32_key(mut conn: PgConnection) {
 
     let rows: Vec<(f32, String)> = r#"
     SELECT id, flavour FROM test_more_like_this_table
-    WHERE test_more_like_this_table @@@ paradedb.more_like_this(
+    WHERE test_more_like_this_table @@@ pdb.more_like_this(
         min_doc_frequency => 0,
         min_term_frequency => 0,
         document_fields => '{"flavour": "banana"}'
@@ -754,7 +754,7 @@ fn more_like_this_f64_key(mut conn: PgConnection) {
     let rows: Vec<(f64, String)> = r#"
     SELECT id, flavour FROM test_more_like_this_table
     WHERE test_more_like_this_table @@@
-    paradedb.more_like_this(
+    pdb.more_like_this(
         min_doc_frequency => 0,
         min_term_frequency => 0,
         document_fields => '{"flavour": "banana"}'
@@ -788,7 +788,7 @@ fn more_like_this_numeric_key(mut conn: PgConnection) {
 
     let rows: Vec<(f64, String)> = r#"
     SELECT CAST(id AS FLOAT8), flavour FROM test_more_like_this_table
-    WHERE test_more_like_this_table @@@ paradedb.more_like_this(
+    WHERE test_more_like_this_table @@@ pdb.more_like_this(
         min_doc_frequency => 0,
         min_term_frequency => 0,
         document_fields => '{"flavour": "banana"}'
@@ -821,7 +821,7 @@ fn more_like_this_date_key(mut conn: PgConnection) {
 
     let rows: Vec<(String, String)> = r#"
     SELECT CAST(id AS TEXT), flavour FROM test_more_like_this_table
-    WHERE test_more_like_this_table @@@  paradedb.more_like_this(
+    WHERE test_more_like_this_table @@@  pdb.more_like_this(
         min_doc_frequency => 0,
         min_term_frequency => 0,
         document_fields => '{"flavour": "banana"}'
@@ -856,7 +856,7 @@ fn more_like_this_time_key(mut conn: PgConnection) {
     let rows: Vec<(String, String)> = r#"
     SELECT CAST(id AS TEXT), flavour FROM test_more_like_this_table
     WHERE test_more_like_this_table @@@
-    paradedb.more_like_this(
+    pdb.more_like_this(
         min_doc_frequency => 0,
         min_term_frequency => 0,
         document_fields => '{"flavour": "banana"}'
@@ -891,7 +891,7 @@ fn more_like_this_timestamp_key(mut conn: PgConnection) {
     let rows: Vec<(String, String)> = r#"
     SELECT CAST(id AS TEXT), flavour FROM test_more_like_this_table
     WHERE test_more_like_this_table @@@
-    paradedb.more_like_this(
+    pdb.more_like_this(
         min_doc_frequency => 0,
         min_term_frequency => 0,
         document_fields => '{"flavour": "banana"}'
@@ -926,7 +926,7 @@ fn more_like_this_timestamptz_key(mut conn: PgConnection) {
     let rows: Vec<(String, String)> = r#"
     SELECT CAST(id AS TEXT), flavour FROM test_more_like_this_table
     WHERE test_more_like_this_table @@@
-    paradedb.more_like_this(
+    pdb.more_like_this(
         min_doc_frequency => 0,
         min_term_frequency => 0,
         document_fields => '{"flavour": "banana"}'
@@ -959,7 +959,7 @@ fn more_like_this_timetz_key(mut conn: PgConnection) {
 
     let rows: Vec<(String, String)> = r#"
     SELECT CAST(id AS TEXT), flavour FROM test_more_like_this_table
-    WHERE test_more_like_this_table @@@ paradedb.more_like_this(
+    WHERE test_more_like_this_table @@@ pdb.more_like_this(
             min_doc_frequency => 0,
             min_term_frequency => 0,
             document_fields => '{"flavour": "banana"}'
